Add files via upload
This is a simple and useful telegram bot. You can use it to change the letters of the Cyrillic alphabet to the Latin alphabet and vice versa.
Attention! To start the bot, you must first get an ID link from Telegram BotFather. You place this link in the desired field specified in the program code. And enjoy it!
Special thanks for - https://github.com/kodchi !